Restructures sorting
