+++ sorting/merge-sort/sort.cc
@@ -1,18 +1,4 @@
-#include "../../main.h"
-
-template <typename BidirIt, typename Compare = std::less<>>
-void merge_sort(BidirIt first, BidirIt last, Compare com = Compare {}){
-    const auto n = std::distance(first, last);
-
-    if(n > 1){
-        const auto middle = std::next(first, n / 2);
-
-        merge_sort(first, middle, com);
-        merge_sort(middle, last, com);
-
-        std::inplace_merge(first, middle, last, com); // the merge stage
-    }
-};
+#include "./sort.h"
 
 int main(){
     std::vector<int> instance(10000000);
