+++ sorting/sort.h
@@ -1,0 +1,8 @@
+#ifndef _ALL_SORTINGS_DEFINITIONS_
+#define _ALL_SORTINGS_DEFINITIONS_
+
+#include "./insertion-sort/sort.h"
+#include "./merge-sort/sort.h"
+#include "./selection-sort/sort.h"
+
+#endif
